pyproject.toml
[tool.poetry]
name = "datatracker"
version = "0.1.0"
description = ""
authors = ["JC Coto <jc@fractalreef.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.0"
Hypercorn = "^0.11.2"
aioredis = "^2.0.0"
httpx = "^0.18.2"
freezegun = "^1.1.0"
pytest-asyncio = "^0.15.1"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
ipdb = "^0.13.9"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
